feat: aplicação de typescript iniciada
