--- output/exemplo,.json
@@ -1,22 +1,0 @@
-[
-  {
-    "Nome": "José",
-    "Stack\r": "FRONT\r"
-  },
-  {
-    "Nome": "Maria",
-    "Stack\r": "BACK\r"
-  },
-  {
-    "Nome": "Lucas",
-    "Stack\r": "FRONT\r"
-  },
-  {
-    "Nome": "Silvia",
-    "Stack\r": "FRONT\r"
-  },
-  {
-    "Nome": "Adriana",
-    "Stack\r": "BACK\r"
-  }
-]
