+++ package.json
@@ -1,11 +1,13 @@
 {
   "devDependencies": {
+    "@types/node": "^18.11.11",
     "typescript": "^4.9.3"
   },
   "name": "csv2json",
   "description": "Primeiro desafio proposto pelo programa GCB Academy: Desenvolver um programa capaz de ler um arquivo de formato .csv e converter o mesmo para formato .json",
   "version": "1.0.0",
   "main": "index.js",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
