+++ src/teste.ts
@@ -1,0 +1,28 @@
+import fs from "fs";
+import path from "path";
+const directory: string = "./input";
+
+function checkDirectory(directory: string) {
+  let directoryExists: boolean = false;
+  if (!fs.existsSync(directory)) {
+    console.error("Directory does not exist: " + directory);
+    return directoryExists;
+  }
+  console.log("Directory exists: " + directory);
+  return (directoryExists = true);
+}
+
+function checkFile(fileWithPath: string) {
+  let fileExists: boolean = false;
+  if (!fs.existsSync(fileWithPath)) {
+    console.error("File does not exist: " + directory);
+    return fileExists;
+  }
+  console.log("File exists: " + directory);
+  return (fileExists = true);
+}
+
+function checkFileExtension(fileWithPath: string, fileExists: boolean) {
+  fileExists = checkFile(fileWithPath);
+  const fileExtension: string = path.extname(fileWithPath)
+}
